+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## NEXT
 ### ✨ Features and improvements
-- None
+- Adds `enableHaloAnimations`, `enableSpaceAnimations`, `disableHaloAnimations`, `disableSpaceAnimations` to Map instance for toggling halo and space animations.
 
 ### 🐛 Bug fixes
 - fixes bug where space would not load correctly when style is a URI.
 - fixes a bug in the UMD module where space crashes when only a color is set.
-- fixes a bug when space: true and no style is passed.
+- fixes a bug when space: true and no style is passed, space becomes unset.
 
 ### Others
 - None