+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # MapTiler SDK Changelog
 
-## NEXT
+## 3.8.0-rc.9
+
 ### ✨ Features and improvements
 - Adds `enableHaloAnimations`, `enableSpaceAnimations`, `disableHaloAnimations`, `disableSpaceAnimations` to Map instance for toggling halo and space animations.
 - Adds custom controls that can be styled however you wish and can do whatever you need, in both declarative (auto-detected) and programmatic way